// src/gen/api/zod/signupBodySchema.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import { z } from 'zod/v4'

export const signupBodySchema = z.object({
  email: z.email().describe('Email address'),
  password: z.string().min(8).max(128),
  firstName: z.string().min(1).max(60),
  lastName: z.string().min(1).max(60),
})